longreport depoly/zillmeterfrontend-longreport-84e007aab0c2/requirements.txt
plotly==5.4.0
matplotlib==3.5.0
pendulum==2.1.2
flask-swagger-ui==3.36.0
kaleido==0.2.1
requests==2.22.0
docx-mailmerge==0.5.0
fpdf==1.7.2
PyPDF2==1.26.0
docx2pdf==0.1.7
apispec==5.1.1
apispec_webframeworks==0.5.2
marshmallow==3.14.1
numpy-financial==1.0.0
boto3==1.20.18